Implement a number of Java.sun.misc.Unsafe.* methods.
Of particular interest is the addition of VMStaticInstance, used to
enable us to return a reference to a static instance, even though a
static instance isn't a heap object (and prevents the underlying
static instance from being GC'd while this reference is live).

This is required for us to support staticFieldBase/staticFieldOffset.

git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@144125 91177308-0d34-0410-b5e6-96231b3b80d8